feat: reauth flow (#167)
* feat: initial working implementation without wallet deletion failure case

* chore: building on implementation

* refactor: refactoring entire reauth implementation to use LoginCoordinator

* chore: building on implementation

* chore: further improvements

* chore: reverting analytics change to WalletCoordinator

* fix: ensuring wallet data is deleted by moving .start for sdk into .start of the coordinator

also auditing access levels of properties and unneeded properties

* chore: updating LoginCoordinator flow

* chore: renaming variable missingPersistentSessionId

* chore: passing only appWindow to LoginCoordinator

* chore: making objc functions private

* test: unit test suite building

* test: amending tests for MainCoordinator and LoginCoordinator

* test: working on structure of LoginCoordinator tests

* test: continuing building on tests

* test: fixing wallet clearWallet test, still failing tests in LoginCoordinator

* refactor: refactored implementation to simplify and singularise the LoginCoordinator flow

* fix: removing testing throw in WalletCoordinator

* chore: adding config for testing implementation without code changes

* refactor: refactoring again, closing app on wallet deletion error screen action

* chore: removing duplicated and unneeded code

* fix: addressing reauth error screen flow bug

* chore: moving clear wallet error feature flag

* test: building test suite

* test: passing test suite

* test: adding welsh strings and adding tests

* test: removing flakey test

* test: adding DeveloperMenuViewControllerTests and fixing view model tests

* chore: adjusting scheme to ensure correct building

* test: flipping conditions in LoginSessionConfigurationTests

* test: amending flakey tests

* test: configuration test values

* test: removing unneeded flakey test file and ensuring consistency in other tests

* test: refining test suite

* chore: removing unneeded argument for fullLogin method

* chore: removing unneeded code and adding notification post to completion handler

* chore: change to missingPersistentSessionId computed variable logic

+++ Sources/Tabs/WalletCoordinator.swift
@@ -10,38 +10,76 @@ final class WalletCoordinator: NSObject,
                                AnyCoordinator,
                                ChildCoordinator,
                                NavigationCoordinator {
-    let window: UIWindow
+    private let window: UIWindow
     let root = UINavigationController()
     weak var parentCoordinator: ParentCoordinator?
-    let analyticsService: AnalyticsService
-    private let secureStoreService: SecureStorable
-    let walletSDK = WalletSDK()
+    private var analyticsCenter: AnalyticsCentral
+    private let userStore: UserStorable
+    private let walletSDK = WalletSDK()
+    private var networkClient = NetworkClient(authenticationProvider: TokenHolder.shared)
     
     init(window: UIWindow,
-         analyticsService: AnalyticsService,
-         secureStoreService: SecureStorable) {
+         analyticsCenter: AnalyticsCentral,
+         userStore: UserStorable) {
         self.window = window
-        self.analyticsService = analyticsService
-        self.secureStoreService = secureStoreService
+        self.analyticsCenter = analyticsCenter
+        self.userStore = userStore
     }
     
     func start() {
         root.tabBarItem = UITabBarItem(title: GDSLocalisedString(stringLiteral: "app_walletTitle").value,
                                        image: UIImage(systemName: "wallet.pass"),
                                        tag: 1)
-    }
-    
-    func updateToken() {
-        let networkClient = NetworkClient(authenticationProvider: TokenHolder.shared)
         walletSDK.start(in: window,
                         with: root,
                         networkClient: networkClient,
-                        analyticsService: analyticsService,
-                        persistentSecureStore: secureStoreService)
-
+                        analyticsService: analyticsCenter.analyticsService,
+                        persistentSecureStore: userStore.openStore)
+        NotificationCenter.default
+            .addObserver(self,
+                         selector: #selector(clearWallet),
+                         name: Notification.Name(.clearWallet),
+                         object: nil)
     }
     
-    func clearWallet() throws {
+    func updateToken() {
+        networkClient = NetworkClient(authenticationProvider: TokenHolder.shared)
+    }
+    
+    func handleUniversalLink(_ url: URL) {
+        walletSDK.deeplink(with: url.absoluteString)
+    }
+    
+    func deleteWalletData() throws {
+        #if DEBUG
+        if AppEnvironment.clearWalletErrorEnabled {
+            throw TokenError.expired
+        }
+        #endif
         try walletSDK.deleteWalletData()
+    }
+    
+    @objc private func clearWallet() {
+        do {
+            try deleteWalletData()
+            userStore.resetPersistentSession()
+            analyticsCenter.analyticsPreferenceStore.hasAcceptedAnalytics = nil
+            let dataDeletionWarningScreen = ErrorPresenter
+                .createDataDeletionWarning(analyticsService: analyticsCenter.analyticsService) { [unowned self] in
+                    window.rootViewController?.presentedViewController?.dismiss(animated: true) {
+                        NotificationCenter.default.post(name: Notification.Name(.returnToIntroScreen), object: nil)
+                    }
+                }
+            dataDeletionWarningScreen.modalPresentationStyle = .overFullScreen
+            window.rootViewController?.presentedViewController?.present(dataDeletionWarningScreen, animated: true)
+        } catch {
+            let unableToLoginErrorScreen = ErrorPresenter
+                .createUnableToLoginError(errorDescription: error.localizedDescription,
+                                          analyticsService: analyticsCenter.analyticsService) {
+                    exit(0)
+                }
+            unableToLoginErrorScreen.modalPresentationStyle = .overFullScreen
+            window.rootViewController?.presentedViewController?.present(unableToLoginErrorScreen, animated: true)
+        }
     }
 }
